+++ grading-admin-service/src/main/java/edu/mines/gradingadmin/config/InterceptorConfig.java
@@ -1,6 +1,7 @@
 package edu.mines.gradingadmin.config;
 
 import edu.mines.gradingadmin.filters.InstructorCourseFilter;
+import edu.mines.gradingadmin.filters.OwnerServicesFilter;
 import edu.mines.gradingadmin.filters.StudentCourseFilter;
 import edu.mines.gradingadmin.managers.SecurityManager;
 import org.springframework.context.annotation.Configuration;
@@ -19,6 +20,8 @@ public class InterceptorConfig implements WebMvcConfigurer {
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(new InstructorCourseFilter(securityManager))
                 .addPathPatterns("/api/instructor/**");
+        registry.addInterceptor(new OwnerServicesFilter(securityManager))
+                .addPathPatterns("/api/owner/**");
         registry.addInterceptor(new StudentCourseFilter(securityManager))
                 .addPathPatterns("/api/student/**");
     }
