+++ grading-admin-service/src/main/java/edu/mines/gradingadmin/filters/OwnerServicesFilter.java
@@ -1,0 +1,47 @@
+package edu.mines.gradingadmin.filters;
+
+import edu.mines.gradingadmin.managers.SecurityManager;
+import edu.mines.gradingadmin.models.enums.CourseRole;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.HandlerMapping;
+
+import java.util.Map;
+import java.util.UUID;
+
+@Slf4j
+public class OwnerServicesFilter implements HandlerInterceptor {
+    private final SecurityManager securityManager;
+
+    public OwnerServicesFilter(SecurityManager securityManager) {
+        this.securityManager = securityManager;
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+        final Map<String, String> pathVariables = (Map<String, String>) request.getAttribute(HandlerMapping.URI_TEMPLATE_VARIABLES_ATTRIBUTE);
+
+        if (!pathVariables.containsKey("course_id")){
+            return true;
+        }
+
+        log.debug("Validating owner access to course for user '{}'", securityManager.getUser().getEmail());
+        String courseId = pathVariables.get("course_id");
+
+        if (courseId.isEmpty()){
+            log.error("Failed to get courseId from protected request!");
+            throw new AccessDeniedException("Failed to get course id from request");
+        }
+
+        if (!securityManager.hasCourseMembership(CourseRole.OWNER, UUID.fromString(courseId))){
+            log.warn("Blocked attempt by '{}' to access course '{}' as an owner!", securityManager.getUser().getEmail(), courseId);
+            throw new AccessDeniedException("Not enrolled in course as an instructor!");
+        }
+
+        return true;
+    }
+}
