+++ grading-admin-service/src/main/java/edu/mines/gradingadmin/services/CourseService.java
@@ -3,22 +3,20 @@ package edu.mines.gradingadmin.services;
 import edu.mines.gradingadmin.config.ExternalServiceConfig;
 import edu.mines.gradingadmin.data.CourseDTO;
 import edu.mines.gradingadmin.data.CourseLateRequestConfigDTO;
-import edu.mines.gradingadmin.data.policyServer.PolicyServerErrorDTO;
 import edu.mines.gradingadmin.events.NewTaskEvent;
 import edu.mines.gradingadmin.managers.IdentityProvider;
 import edu.mines.gradingadmin.managers.ImpersonationManager;
 import edu.mines.gradingadmin.models.*;
+import edu.mines.gradingadmin.models.enums.CourseRole;
 import edu.mines.gradingadmin.models.tasks.ScheduledTaskDef;
 import edu.mines.gradingadmin.models.tasks.CourseSyncTaskDef;
 import edu.mines.gradingadmin.repositories.*;
 import edu.mines.gradingadmin.services.external.CanvasService;
-import edu.mines.gradingadmin.services.external.PolicyServerService;
 import edu.mines.gradingadmin.services.external.S3Service;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.net.URI;
@@ -37,6 +35,7 @@ public class CourseService {
     private final S3Service s3Service;
     private final PolicyRepo policyRepo;
     private final UserService userService;
+
     public CourseService(CourseRepo courseRepo, CourseLateRequestConfigRepo lateRequestConfigRepo, GradescopeConfigRepo gradescopeConfigRepo, ScheduledTaskRepo<CourseSyncTaskDef> taskRepo,
 
                          ApplicationEventPublisher eventPublisher, ImpersonationManager impersonationManager,
@@ -54,35 +53,36 @@ public class CourseService {
         this.userService = userService;
     }
 
-    public List<Course> getCourses(boolean enabled) {
-            if (enabled) {
-                return courseRepo.getAll(enabled);
-            }
-            return courseRepo.getAll();
+    public List<Course> getAllCourses(boolean enabled) {
+        if (enabled) {
+            return courseRepo.getAll(true);
+        }
+        return courseRepo.getAll();
     }
+
 
     public List<Course> getCoursesStudent(User user){
-        List<CourseMember> memberships = userService.getCourseMemberships(user.getCwid());
-        List<Course> existingCourses = courseRepo.getAll(true);
-        List<Course> studentEnrolledCourses = new ArrayList<>();;
-        for (CourseMember courseMember : memberships) {
-            for(Course course: existingCourses) {
-                if (courseMember.getCourse().equals(course)) {
-                    studentEnrolledCourses.add(course);
-                }
-            }
-        }
-        return studentEnrolledCourses;
+        return getCoursesByRole(user, CourseRole.STUDENT, true);
     }
 
-    public Optional<Course> getCourse(UUID courseId){
+    public List<Course> getCoursesByRole(User user, CourseRole courseRole, Boolean enabled) {
+        List<CourseMember> memberships = userService.getCourseMemberships(user.getCwid());
+
+        return memberships.stream().filter(m -> m.getRole().equals(courseRole))
+                .map(CourseMember::getCourse)
+                // only filter on enabled if boolean is non-null - this is why it has the object instead of primitive
+                .filter(c -> enabled == null || c.isEnabled() == enabled)
+                .toList();
+    }
+
+    public Optional<Course> getCourse(UUID courseId) {
         return courseRepo.getById(courseId);
     }
 
     public Optional<Course> updateCourse(String courseId, CourseDTO courseDTO) {
         Optional<Course> course = getCourse(UUID.fromString(courseId));
 
-        if(course.isEmpty()) {
+        if (course.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Course does not exist");
         }
 
@@ -101,8 +101,8 @@ public class CourseService {
         }
 
         GradescopeConfig gsConfig = course.get().getGradescopeConfig();
-        if(gsConfig != null) {
-            if(courseDTO.getGradescopeId() != null) {
+        if (gsConfig != null) {
+            if (courseDTO.getGradescopeId() != null) {
                 gsConfig.setGradescopeId(courseDTO.getGradescopeId().toString());
             }
             gsConfig.setEnabled(courseDTO.getEnabled());
@@ -112,7 +112,7 @@ public class CourseService {
         return Optional.of(courseRepo.save(course.get()));
     }
 
-    public void syncCourseTask(CourseSyncTaskDef task){
+    public void syncCourseTask(CourseSyncTaskDef task) {
         IdentityProvider user = impersonationManager.impersonateUser(task.getCreatedByUser());
         List<edu.ksu.canvas.model.Course> availableCourses =
                 canvasService.asUser(user).getAllAvailableCourses()
@@ -134,7 +134,7 @@ public class CourseService {
 
         Optional<Course> toUpdate = courseRepo.getById(task.getCourseToSync());
 
-        if (toUpdate.isEmpty()){
+        if (toUpdate.isEmpty()) {
             log.warn("Course to sync not found!");
             throw new RuntimeException(String.format("Failed to find course: '%s'", task.getCourseToSync()));
         }
@@ -149,8 +149,8 @@ public class CourseService {
     }
 
     public Optional<ScheduledTaskDef> syncCourseWithCanvas(User actingUser, UUID courseId, long canvasId,
-                                                             boolean overwriteName, boolean overwriteCode) {
-        if (!courseRepo.existsById(courseId)){
+                                                           boolean overwriteName, boolean overwriteCode) {
+        if (!courseRepo.existsById(courseId)) {
             log.warn("Course '{}' has not been created!", courseId);
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Course does not exist");
         }
@@ -174,7 +174,7 @@ public class CourseService {
 
     public void enableCourse(UUID courseId) {
         Optional<Course> course = courseRepo.findById(courseId);
-        if(course.isPresent()) {
+        if (course.isPresent()) {
             course.get().setEnabled(true);
             courseRepo.save(course.get());
         }
@@ -182,13 +182,13 @@ public class CourseService {
 
     public void disableCourse(UUID courseId) {
         Optional<Course> course = courseRepo.findById(courseId);
-        if(course.isPresent()) {
+        if (course.isPresent()) {
             course.get().setEnabled(false);
             courseRepo.save(course.get());
         }
     }
 
-    public Optional<Course> createNewCourse(CourseDTO courseDTO){
+    public Optional<Course> createNewCourse(CourseDTO courseDTO) {
         Course newCourse = new Course();
         newCourse.setName(courseDTO.getName());
         newCourse.setCode(courseDTO.getCode());
@@ -198,7 +198,7 @@ public class CourseService {
         Optional<CourseLateRequestConfig> lateRequestConfig = createCourseLateRequestConfig(courseDTO.getLateRequestConfig());
 
         ExternalServiceConfig externalServiceConfig = new ExternalServiceConfig();
-        if(courseDTO.getGradescopeId() != null) {
+        if (courseDTO.getGradescopeId() != null) {
             ExternalServiceConfig.GradescopeConfig config = externalServiceConfig.configureGradescope(true, URI.create("https://www.gradescope.com/courses/" + courseDTO.getGradescopeId()));
 
             GradescopeConfig gsConfig = new GradescopeConfig();
@@ -215,7 +215,7 @@ public class CourseService {
 
         Optional<String> bucketName = s3Service.createNewBucketForCourse(newCourse.getId());
 
-        if (bucketName.isEmpty()){
+        if (bucketName.isEmpty()) {
             log.warn("Failed to create S3 bucket!");
         }
 
@@ -223,7 +223,7 @@ public class CourseService {
     }
 
     private Optional<CourseLateRequestConfig> createCourseLateRequestConfig(CourseLateRequestConfigDTO dto) {
-        if (dto == null){
+        if (dto == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Course late request does not exist");
         }
         CourseLateRequestConfig lateRequestConfig = new CourseLateRequestConfig();
